requirements.txt
langchain==0.3.24
langchain-openai==0.3.14
langchain-community==0.3.22
langchain-core==0.3.55
langgraph==0.3.33
langsmith==0.3.33
streamlit==1.44.1
qdrant-client==1.13.3
exa-py==1.12.1
python-dotenv==1.1.0
